epic-lab: fix an import path issue
